// frontend/config/types.ts
export type Status = "idle" | "pending" | "error" | "success"
export type Vote = 0n | 1n | 2n  // = against, for, abstain  
// 'string | number | bigint | boolean | ByteArray 
export type OrganizationType = 'Powers 101' | 'Bridging Off-Chain Governance' | 'Grants Manager' | 'Split Governance' | 'Packaged Upgrades' | 'Single Upgrades'
export type InputType = boolean | string | number | bigint | `0x${string}`
export type DataType = "uint8" | "uint16" | "uint32" | "uint48" | "uint64" | "uint128" | "uint256" | "address" | "bytes" | "string" | "bytes32" | "bool" |
                       "uint8[]" | "uint16[]" | "uint32[]" | "uint48[]" | "uint64[]" | "uint128[]" | "uint256[]" | "address[]" | "bytes[]" | "string[]" | "bytes32[]" | "bool[]" | "unsupported" | "empty" 
export type LawSimulation = [
      bigint, 
      `0x${string}`[], 
      bigint[], 
      `0x${string}`[], 
      `0x${string}`
]

export type LawExecutions = {
  powers: `0x${string}`, 
  config: `0x${string}`, 
  actionsIds: bigint[], 
  executions: bigint[] 
}
// ...
export type Conditions = {
  allowedRole: bigint; 
  delayExecution: bigint; 
  needNotCompleted: bigint;
  needCompleted: bigint;
  readStateFrom: bigint;
  quorum: bigint; 
  succeedAt: bigint; 
  throttleExecution: bigint;
  votingPeriod: bigint;
}

type Args = {
  nonce: bigint;
  description: string;
  caller: `0x${string}`;
  lawCalldata: `0x${string}`;
  targetLaw: `0x${string}`;
}

export type LogExtended = Log & 
  {args: Args}

export type Execution = {
  log: LogExtended; 
  blocksData?: GetBlockReturnType
}

export type Law = {
  powers: `0x${string}`;
  lawAddress: `0x${string}`;
  lawHash: `0x${string}`;
  index: bigint;
  nameDescription?: string;
  conditions?: Conditions;
  config?: `0x${string}`;
  inputParams?: `0x${string}`; 
  params ?: {varName: string, dataType: DataType}[]; 
  executions?: LawExecutions; 
  active: boolean;
}